requirements.txt
streamlit==1.29.0
openai==2.8.1
python-dotenv==1.0.0
pydantic==2.5.2
regex==2023.10.3
groq==0.36.0
huggingface-hub==1.1.6
requests==2.32.5